requirements.txt
nose==1.3.7
spacy==2.0.8
ufal.udpipe==1.2.0.1
nltk==3.2.5
